correction d'un bug sur delete de contributorviewset.py

+++ projects/views/contributorviewset.py
@@ -83,23 +83,25 @@ class ContributorViewset(MultipleSerializerMixin, ModelViewSet):
 
     def destroy(self, request, *args, **kwargs):
         project_id = self.kwargs['project_id']
-        project = get_object_or_404(
-            Project.objects.filter(id=project_id)
-        )
-        if project.author_user_id == self.request.user:
-            tmp_serializer = self.get_serializer(data=request.data)
-            user = get_object_or_404(
-                User.objects.filter(
-                    username=tmp_serializer.initial_data['user_id'])
-            )
-            Contributor.objects.filter(
-                    project_id=project_id,
-                    user_id=user
-                ).delete()
-            return Response(
-                {'message': 'The user has been delete from project'},
-                status=status.HTTP_200_OK
-            )
+
+        if Project.objects.filter(id=project_id).exists():
+            project = Project.objects.get(id=project_id)
+            if project.author_user_id == self.request.user:
+                user = get_object_or_404(
+                    User.objects.filter(
+                        pk=self.kwargs['pk'])
+                )
+                Contributor.objects.filter(
+                        project_id=project_id,
+                        user_id=user
+                    ).delete()
+                return Response(
+                    {'message': 'The user has been delete from project'},
+                    status=status.HTTP_200_OK
+                )
+            else:
+                return Response({'message': 'Forbidden action'},
+                                status=status.HTTP_403_FORBIDDEN)
         else:
-            return Response({'message': 'Forbidden action'},
-                            status=status.HTTP_403_FORBIDDEN)
+            return Response({'project': "this project doesn't exists"},
+                            status=status.HTTP_400_BAD_REQUEST)
